requirements.txt
fastapi
uvicorn
openai
python-dotenv
upstash-redis
supabase
slowapi
pydantic
pypdf
python-multipart

